apply braces method overloaded

+++ src/main/java/com/sv/core/Utils.java
@@ -315,7 +315,7 @@ public class Utils {
         return applyBraces(s+"");
     }
 
-    public static String applyBraces(int s) {
+    public static String applyBraces(long s) {
         return applyBraces(s+"");
     }
